requirements.txt
py-clob-client>=0.5.0
web3>=6.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
aiohttp>=3.9.0
websockets>=12.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
vectorbt>=0.26.0
